Renamed fs module to io
